+++ patzilla-ui/access/epo-ops.js
@@ -533,7 +533,7 @@ _.extend(OpsExchangeDocument.prototype, {
         },
 
         get_classification_schemes: function() {
-            return ['IPC', 'IPC-R', 'CPC', 'UC', 'FI', 'FTERM'];
+            return ['IPC', 'IPC-R', 'CPC', 'CPCNO', 'UC', 'FI', 'FTERM'];
         },
 
         get_classifications: function(links) {
@@ -600,7 +600,7 @@ _.extend(OpsExchangeDocument.prototype, {
                     _.defaults(classifications, defaults);
 
                     var entry;
-                    if (scheme == 'CPC') {
+                    if (scheme == 'CPC' || scheme == 'CPCNO') {
                         var entry_parts = [];
                         _(cpc_fieldnames).each(function(cpc_fieldname) {
                             if (cpc_fieldname == '/') {
@@ -635,7 +635,7 @@ _.extend(OpsExchangeDocument.prototype, {
 
             if (links) {
                 _.each(classifications, function(value, key) {
-                    if (key == 'CPC') {
+                    if (key == 'CPC' || key == 'CPCNO') {
                         classifications[key] = _this.enrich_links(value, 'cpc', quotate);
                     }
                 });