+++ CHANGES.rst
@@ -4,9 +4,18 @@ IP Navigator CHANGES
 
 development
 ===========
+
+2017-10-13 0.141.0
+==================
 - Enable display of "CPCNO" classifications
-- Fix/improve fulltext display for e.g. KR20170103976A
-- IFI Claims fulltext sections “industrial-applicability” and “reference-signs-list” were missing
+- IFI Claims fulltext
+
+    - Add proper escaping and newline replacement for description, e.g. KR20170103976A
+    - Description sections “industrial-applicability” and “reference-signs-list” were missing, e.g. KR20170103976A
+    - Update claims structure, e.g. JP2017128728A
+    - Update "description-of-drawings" description section, e.g. JP2017128728A
+    - Parse "chemistry" and other figref nodes in "summary-of-invention.tech-solution" section, e.g. JP2017128728A
+    - Description section "description-of-embodiments.embodiments-example" was missing, e.g. JP2017128728A
 
 2017-10-12 0.140.2
 ==================